requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart>=0.0.18

gradio==5.8.0
pillow==11.0.0

pymupdf==1.24.14

numpy==1.26.4
faiss-cpu==1.8.0.post1

torch
transformers
accelerate
